feat: Add Consumer Logic
