+++ higgins-cli/src/main.rs
@@ -7,6 +7,10 @@ use tokio::{
     net::TcpStream,
 };
 
+use crate::ping::handle_ping_cmd;
+
+mod ping;
+
 /// Simple program to greet a person
 #[derive(Parser, Debug)]
 #[command(version, about, long_about = None)]
@@ -26,6 +30,11 @@ enum Commands {
         #[arg(long, require_equals = true)]
         file_name: String,
     },
+    CreateConsumer {
+        #[arg(long, require_equals = true)]
+        topic: String,
+        // partitions?
+    },
 }
 
 #[tokio::main]
@@ -44,101 +53,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     match args.command {
         Some(cmd) if matches!(cmd, Commands::Ping {}) => {
-            // Send a ping command to the server;
-
-            let mut read_buf = BytesMut::zeroed(20);
-
-            let mut write_buf = BytesMut::new();
-
-            let ping = Ping::default();
-
-            Message {
-                r#type: Type::Ping as i32,
-                ping: Some(ping),
-                ..Default::default()
-            }
-            .encode(&mut write_buf)
-            .unwrap();
-
-            tracing::info!("Writing: {:#?}", write_buf);
-
-            let result = socket.write_all(&write_buf).await.unwrap();
-
-            let n = socket.read(&mut read_buf).await.unwrap();
-
-            tracing::info!("Received: {:#?}", socket.peer_addr());
-
-            tracing::info!("Reading: {:#?}", read_buf.clone().to_vec());
-
-            let slice = &read_buf[0..n];
-
-            if n == 0 {
-                tracing::info!("No bytes read, continuing.");
-                return Ok(());
-            }
-
-            let message = Message::decode(slice).unwrap();
-
-            match Type::try_from(message.r#type).unwrap() {
-                Type::Ping => {
-                    let mut result = BytesMut::new();
-
-                    let pong = Pong::default();
-
-                    Message {
-                        r#type: Type::Pong as i32,
-                        pong: Some(pong),
-                        ..Default::default()
-                    }
-                    .encode(&mut result)
-                    .unwrap();
-
-                    socket.write(&result).await.unwrap();
-                }
-                Type::Consumerequest => todo!(),
-                Type::Consumeresponse => todo!(),
-                Type::Producerequest => todo!(),
-                Type::Produceresponse => todo!(),
-                Type::Metadatarequest => todo!(),
-                Type::Metadataesponse => todo!(),
-                Type::Pong => {
-                    tracing::info!("Received Pong!")
-                }
-            }
-
-            let result = socket.write_all(&write_buf).await.unwrap();
-            let n = socket.read(&mut read_buf).await.unwrap();
-
-            let slice = &read_buf[0..n];
-
-            let message = Message::decode(slice).unwrap();
-
-            match Type::try_from(message.r#type).unwrap() {
-                Type::Ping => {
-                    let mut result = BytesMut::new();
-
-                    let pong = Pong::default();
-
-                    Message {
-                        r#type: Type::Pong as i32,
-                        pong: Some(pong),
-                        ..Default::default()
-                    }
-                    .encode(&mut result)
-                    .unwrap();
-
-                    socket.write(&result).await.unwrap();
-                }
-                Type::Consumerequest => todo!(),
-                Type::Consumeresponse => todo!(),
-                Type::Producerequest => todo!(),
-                Type::Produceresponse => todo!(),
-                Type::Metadatarequest => todo!(),
-                Type::Metadataesponse => todo!(),
-                Type::Pong => {
-                    tracing::info!("Received Pong!")
-                }
-            }
+            handle_ping_cmd(&mut socket).await;
         }
         Some(cmd)
             if matches!(
